Added some more arguments
All the new arguments are set by default to the setting used in the Proc B paper (so, the function can still be called exactly the same way as done in the Markdown in this directory). 'return_plot_table' allows returning the final table used by the function to for plotting, as well as returning the jittered coordinates of the points. 'line_colours' allows setting the colours of the outlines of the circles, following the same logic as 'dot_colours'.